styel: Page7 스타일 재정의
